Change OkHttpConnector to OkHttpGitHubConnector
To preserve backward compatibility, OkHttpConnector must continue to be based on HttpConnector.
