+++ pom.xml
@@ -153,12 +153,17 @@
                       </limit>
                     </limits>
                     <excludes>
+                      <!-- Code implemented externally -->
+                      <exclude>org.kohsuke.github.extras.okhttp3.ObsoleteUrlFactory.**</exclude>
+                      <exclude>org.kohsuke.github.extras.okhttp3.ObsoleteUrlFactory</exclude>
+
                       <!-- Sample only -->
                       <exclude>org.kohsuke.github.example.*</exclude>
 
 
                       <!-- Deprecated -->
-                      <exclude>org.kohsuke.github.extras.OkHttpConnector</exclude>
+                      <exclude>org.kohsuke.github.extras.okhttp3.OkHttpConnector</exclude>
+                      <exclude>org.kohsuke.github.extras.OkHttp3Connector</exclude>
                       <exclude>org.kohsuke.github.extras.OkHttp3Connector</exclude>
                       <exclude>org.kohsuke.github.EnforcementLevel</exclude>
                       <exclude>org.kohsuke.github.GHPerson.1</exclude>
