+++ src/test/java/org/kohsuke/github/GHRateLimitTest.java
@@ -16,7 +16,7 @@ import static org.hamcrest.Matchers.*;
 import static org.hamcrest.core.IsInstanceOf.instanceOf;
 
 /**
- * Test showing the behavior of OkHttpConnector with and without cache.
+ * Test showing the behavior of OkHttpGitHubConnector with and without cache.
  * <p>
  * Key take aways:
  *
